Fix issue going to next day of games
