+++ data/schedule.py
@@ -1,4 +1,5 @@
 import time
+from datetime import datetime, timedelta
 
 import statsapi
 
@@ -8,13 +9,13 @@ from data import status
 from data.game import Game
 from data.update import UpdateStatus
 
-GAMES_REFRESH_RATE = 5 * 60
+GAMES_REFRESH_RATE = 6 * 60
 
 
 class Schedule:
-    def __init__(self, date, config):
-        self.date = date.strftime("%Y-%m-%d")
+    def __init__(self, config):
         self.config = config
+        self.date = self.__parse_today()
         self.starttime = time.time()
         self.current_idx = 0
         self.preferred_over = False
@@ -24,12 +25,25 @@ class Schedule:
         self._games = []
         self.update(True)
 
+    def __parse_today(self):
+        if self.config.demo_date:
+            today = datetime.strptime(self.config.demo_date, "%Y-%m-%d")
+        else:
+            today = datetime.today()
+            end_of_day = datetime.strptime(self.config.end_of_day, "%H:%M").replace(
+                year=today.year, month=today.month, day=today.day
+            )
+            if end_of_day > datetime.now():
+                today -= timedelta(days=1)
+        return today
+
     def update(self, force=False) -> UpdateStatus:
         if force or self.__should_update():
+            self.date = self.__parse_today()
             debug.log("Updating schedule for %s", self.date)
             self.starttime = time.time()
             try:
-                self.__all_games = statsapi.schedule(self.date)
+                self.__all_games = statsapi.schedule(self.date.strftime("%Y-%m-%d"))
             except:
                 debug.exception("Networking error while refreshing schedule")
                 return UpdateStatus.FAIL
